requirements.txt
# Core API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-multipart==0.0.6

# ML and Data Science dependencies
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
joblib==1.5.0
tensorflow-cpu==2.16.1  # Using latest stable version compatible with Python 3.12
keras>=2.16.1
shap==0.47.2

# Utility dependencies
typing_extensions==4.13.2
protobuf>=3.20.3,<5.0.0dev
